NPC refactored
NPC is no longer generated by this plugin. Instead players are encouraged to create the NPC using Citizens and then input the ID into the config. This allows for multiple NPC's.

Item Preset introduction

When a new shipment is generated, a preset can now be chosen to apply items automatically.